Some progress made toward rendering.
We are skipping over the 2D part and going straight to 3D. My plan is
to adapt the program from my Computer Grapics class in college. That
one was made in C++ and renders polygons, so not quite what I am
going for.

Eiter way, progress made. The Matrix class in Physics_Sim.Graphics will
almost definitely be deleted. I put it together before realizing the Xna
framework has the exact  thing I was making. Still, the file is there.