Started to render from Coordinate System.
I am not sure that is the most descriptive title, but that is what this
part is for. The Matrix class is gone, it was of no use. Vector has been
fixed up and is in use.

CoordinateSystem is now using render instead of getGraphic, that is
a start to using 3D space. There is currently a point at the origin that
it is trying to render. It may be doing that correctly, but without
drawing the axis I really dont know. That is for future me to test, I am
out of time to work on this now.