Testing out a different grid.
This update was an attempt at getting a more usable grid implemented.
The new grid does look better, and uses guide lines with the aability to
draw/not draw by quadrent. The issue is, the vector will not overlay it
correctly.

My best guess this that each quadrent will need to control the drawing
of the vectors inside of it, but I don't know if I am willing to
implement that. I may, I may not. That's why it's a branch.