Made grid adjustable.
Alright, so we still are not putting vectors on the grid. The grid is
much more adjustable now. You can change the size of each quadrent,
change what quadrent is shown, and change wich plane has a grid
displayed on it.

So in other words, it is better but still not usable. In contrast to the
original grid that is bad but usable. Cool.