Vectors sort of added.
Alright, it's not super clean yet - but we are displaying our vectors.
It looks pretty nice on the new grid as well. Seeing how well this
branch has gone, it is just going to overwrite the master. That was not
the original plan, otherwise I would not have branched.

The point of building the program is showing through though, as I can
graphicly show that I do not know how to add vectors.